add HandlerRegistry
